+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/ext/TestDropwizardAppExtension.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.platform.commons.support.AnnotationSupport;
 import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
+import ru.vyarus.dropwizard.guice.test.client.ApacheTestClientFactory;
 import ru.vyarus.dropwizard.guice.test.jupiter.TestDropwizardApp;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.TestEnvironmentSetup;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.conf.ExtensionBuilder;
@@ -364,7 +365,11 @@ public class TestDropwizardAppExtension extends GuiceyExtensionsSupport {
             res.tracker.debug = ann.debug();
             res.reuseApp = ann.reuseApplication();
             res.defaultExtensionsEnabled = ann.useDefaultExtensions();
-            res.clientFactory(ann.clientFactory());
+            if (ann.useApacheClient()) {
+                res.clientFactory(ApacheTestClientFactory.class);
+            } else {
+                res.clientFactory(ann.clientFactory());
+            }
             return res;
         }
     }